requirements.txt
click==8.3.0
colorama==0.4.6
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
PyYAML==6.0.3
rich==14.2.0
shellingham==1.5.4
typer==0.19.2
typing_extensions==4.15.0
networkx>=3.0
matplotlib>=3.7.0
